start recovery of files
